+++ quiz/src/main/java/com/epam/spring/homework3/quiz/service/GameService.java
@@ -3,7 +3,7 @@ package com.epam.spring.homework3.quiz.service;
 import com.epam.spring.homework3.quiz.controller.dto.QuizDto;
 
 public interface GameService {
-    QuizDto startGame(Integer id);
+    QuizDto startGame(Long id);
 
-    String checkResultOfGame(QuizDto quizDto, Integer id, String userName);
+    String checkResultOfGame(QuizDto quizDto, Long id, String userName);
 }
